move ContractCallReward to platform folder
