+++ src/components/[guild]/RoleCard/RoleCard.tsx
@@ -18,6 +18,7 @@ import GuildLogo from "components/common/GuildLogo"
 import dynamic from "next/dynamic"
 import { memo, useEffect, useRef } from "react"
 import { PlatformType, Role, Visibility as VisibilityType } from "types"
+import { ContractCallRewardIcon } from "../../../platforms/ContractCall/ContractCallReward"
 import useAccess from "../hooks/useAccess"
 import useGuild from "../hooks/useGuild"
 import useGuildPermission from "../hooks/useGuildPermission"
@@ -28,7 +29,6 @@ import AccessIndicator from "./components/AccessIndicator"
 import HiddenRewards from "./components/HiddenRewards"
 import MemberCount from "./components/MemberCount"
 import Reward, { RewardIcon } from "./components/Reward"
-import { ContractCallRewardIcon } from "./components/Reward/components/ContractCallReward"
 import RoleDescription from "./components/RoleDescription"
 
 type Props = {
