--- src/components/[guild]/RoleCard/components/Reward/index.ts
@@ -1,4 +1,0 @@
-import Reward, { RewardDisplay, RewardIcon } from "./Reward"
-
-export default Reward
-export { RewardDisplay, RewardIcon }
